+++ projects/playground/src/app/pages/overlay-test-pages/toast-test-page/toast-test-page.component.ts
@@ -16,7 +16,7 @@ export class ToastTestPageComponent {
   protected ToastTestPageComponent = ToastTestPageComponent
 
   private readonly quangTranslationService = inject(QuangTranslationService)
-  private readonly customToast = viewChild<TemplateRef<any>>('customToast')
+  private readonly customToast = viewChild<TemplateRef<unknown>>('customToast')
   private readonly quangToast = inject(QuangToastService)
   private readonly testComponent = viewChild('testComponent')