+++ projects/playground/src/app/shared/directives/source-code.directive.ts
@@ -1,15 +1,14 @@
-import { Directive, ElementRef, Renderer2 } from '@angular/core'
+import { Directive, ElementRef, Renderer2, inject } from '@angular/core'
 
 @Directive({
-  selector: '[playgroundSourceCode]'
+  selector: '[playgroundSourceCode]',
 })
-export class SourceCodeDirective{
+export class SourceCodeDirective {
+  private readonly el = inject<ElementRef<HTMLElement>>(ElementRef)
+  private readonly renderer = inject(Renderer2)
 
-  constructor(private readonly el: ElementRef, private readonly renderer: Renderer2) { // Get the outerHTML of the element before Angular processes it
-    const html = this.el.nativeElement.outerHTML;
-
-    // Store the raw HTML as a data attribute
-    this.renderer.setAttribute(this.el.nativeElement, 'data-source', html);
+  constructor() {
+    const html = this.el.nativeElement.outerHTML
+    this.renderer.setAttribute(this.el.nativeElement, 'data-source', html)
   }
-
 }